git cherry-pick 944ce8d and merge conflicts
