+++ ios/FluentUI/Avatar/AvatarGroupView.swift
