+++ ios/FluentUI/Avatar/Presence.swift
