+++ ios/FluentUI/Navigation/Shy Header/ShyHeaderController.swift
@@ -77,7 +77,9 @@ class ShyHeaderController: UIViewController {
         }
 
         loadViewIfNeeded()
-        addChildController(contentViewController, containingViewIn: contentContainerView)
+        addChild(contentViewController)
+        contentContainerView.addSubview(contentViewController.view)
+        contentViewController.didMove(toParent: self)
         contentViewController.view.fitIntoSuperview(usingConstraints: true)
 
         contentScrollViewObservation = contentViewController.navigationItem.observe(\.contentScrollView, options: [.new]) { [weak self] (_, change) in
@@ -241,7 +243,7 @@ class ShyHeaderController: UIViewController {
 
         // if the originator is a VC, make sure it belongs to this heirarchy
         if let originatorVC = expansionRequestOriginator as? UIViewController {
-            guard originatorVC == contentViewController || contentViewController.isAncestor(ofViewController: originatorVC) else {
+            guard originatorVC == contentViewController || isViewControllerAncestor(contentViewController, ofViewController: originatorVC) else {
                 return false
             }
         }
@@ -483,6 +485,18 @@ class ShyHeaderController: UIViewController {
                        animations: {
                         self.updateHeader(with: 0.0, expanding: false)
         }, completion: nil)
+    }
+
+    private func isViewControllerAncestor(_ ancestorViewController: UIViewController, ofViewController descendantViewController: UIViewController) -> Bool {
+        for child in ancestorViewController.children {
+            if child == descendantViewController {
+                return true
+            }
+            if isViewControllerAncestor(child, ofViewController: descendantViewController) {
+                return true
+            }
+        }
+        return false
     }
 
     /// Sets the animation progress, using progress as a fraction not a percent
